+++ .changeset/cruel-readers-peel.md
@@ -1,0 +1,5 @@
+---
+"@turnkey/react-wallet-kit": patch
+---
+
+Fixed type re-exports from `@turnkey/core`
