+++ packages/core/src/__types__/index.ts
@@ -10,6 +10,7 @@
 
 export * from "./auth";
 export * from "./config";
+export * from "./enums";
 export * from "./error";
 export * from "./export";
 export * from "./external-wallets";
