+++ packages/core/src/index.ts
@@ -19,6 +19,7 @@ export { type TurnkeySDKClientBase } from "./__generated__/sdk-client-base";
 // Export all types and values from __types__/
 export * from "./__types__/auth";
 export * from "./__types__/config";
+export * from "./__types__/enums";
 export * from "./__types__/error";
 export * from "./__types__/export";
 export * from "./__types__/external-wallets";
